Convert TimeEventFilter into hook
